Move MainView into Views directory
