+++ WinForms/Program.cs
@@ -11,7 +11,7 @@ namespace nardnob.InputTracker.WinForms
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new MainView());
+            Application.Run(new Views.MainView());
         }
     }
 }
