+++ WinForms/Views/MainView.resx
